RS5 - Mikroservisna arhitektura/examples/e-commerce-app/order-service/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
frozenlist==1.8.0
idna==3.11
infobip-api-python-client==5.1.1
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.0.7
yarl==1.22.0
